+++ src/main/java/app/controller/ServerBrowserController.java
@@ -9,7 +9,6 @@ class ServerBrowserController extends Controller {
 
   private static final String FXMLPATH = "view/serverbrowser.fxml";
   private static final String USERNAME_ERROR = "Please choose a username.";
-  private static final String ALERT_HEADER = "Username error.";
 
   @FXML
   TextField username;
@@ -21,7 +20,7 @@ class ServerBrowserController extends Controller {
   @FXML
   private void joinGame() {
     if (username.getText().equals("")) {
-      displayUsernameError();
+      displayError(USERNAME_ERROR);
     } else {
       new GameLobbyController(stage);
     }
@@ -30,19 +29,9 @@ class ServerBrowserController extends Controller {
   @FXML
   private void switchToCreateScreen() {
     if (username.getText().equals("")) {
-      displayUsernameError();
+      displayError(USERNAME_ERROR);
     } else {
       new CreateController(stage);
     }
   }
-
-  private void displayUsernameError() {
-    Alert alert = new Alert(Alert.AlertType.WARNING);
-    alert.setTitle(ALERT_HEADER);
-    alert.setContentText(USERNAME_ERROR);
-    alert.setHeaderText(null);
-    alert.showAndWait();
-  }
-
-
 }