+++ src/main/java/client/Main.java
@@ -2,6 +2,9 @@ package client;
 
 import java.io.IOException;
 import java.net.InetAddress;
+import java.util.List;
+import protocol.LobbyModel;
+import protocol.RaceModel;
 
 /**
  * Temporary Client main for testing.
@@ -20,11 +23,18 @@ public class Main implements ClientObserver {
       Client client = new ClientImpl(addr, 8080);
       client.subscribe(m);
       client.registerUser("Cooler Mensch");
+      client.requestLobbies();
       client.newGame("some userId");
+      client.requestLobbies();
       client.startGame();
     } catch (IOException e) {
       System.out.print(e.getMessage());
     }
+  }
+
+  @Override
+  public void registered(String userId) {
+    System.out.println("Registered with userId: " + userId);
   }
 
   @Override
@@ -33,22 +43,23 @@ public class Main implements ClientObserver {
   }
 
   @Override
-  public void joinedGame(String gameId, boolean isRunning) {
-    System.out.println("Joined Game " + gameId);
+  public void gameStarting(RaceModel race) {
+    System.out.println("Game starting with text: " + race.textToType);
   }
 
   @Override
-  public void gameStarting(String text) {
-    System.out.println("Game starting with text: " + text);
+  public void receivedLobbyUpdate(LobbyModel lobby) {
+    System.out.println("Received Lobby update:");
+    System.out.println(lobby.players);
+    System.out.println(lobby.id);
+    System.out.println(lobby.isRunning);
   }
 
   @Override
-  public void playerJoined(String name) {
-    System.out.println("Player joined: " + name);
+  public void receivedOpenLobbies(List<LobbyModel> lobbies) {
+    System.out.println("Open lobbies:");
+    System.out.println(lobbies);
   }
 
-  @Override
-  public void playerLeft(String name) {
-    System.out.println("Player Left: " + name);
-  }
+
 }